requirements.txt
fastapi
uvicorn
langchain
langgraph
langsmith
openai
pypdf
tiktoken
python-dotenv
yfinance
alpha_vantage
faiss-cpu
sentence-transformers
pandas
requests
